refactored userController.go's functions that manage auth to new authController.go
